+++ src/program/resolve/index.rs
@@ -721,6 +721,7 @@ mod tests {
             abi_event_kind: None,
             param_sig_key: String::new(),
             return_type: None,
+            return_type_id: None,
         }
     }
 
@@ -749,6 +750,7 @@ mod tests {
             abi_event_kind: None,
             param_sig_key: String::new(),
             return_type: None,
+            return_type_id: None,
         }
     }
 
@@ -778,6 +780,7 @@ mod tests {
             abi_event_kind: None,
             param_sig_key: String::new(),
             return_type: None,
+            return_type_id: None,
         }
     }